+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a view displaying your daily log from the report page. (Currently only available as an experiment).
 
+### Changed
+
+- Now all objectives display their duration in their division.
+
 ### Fixed
 
 - Fixed a performance issue in the homepage caused by notification validation looping itself. The "Loading..." screen became slightly longer, in exchange.